sched/tune: Consider coordinate axis and origin
We use SchedTune PE filter to compare the energy and capacity
difference between two CPUs; if the condition is:
(nrg_delta, cap_delta) = (0, 0) then this is origin of PE filter's
coordinate, that means the two CPUs has no difference from either the
power or performance perspective for waken task; usually we prefer
to keep the task to run on previous CPU and avoid to migrate to other
new CPU.  But PE filter misses to handle origin and the function
__schedtune_accept_deltas() cannot find proper 'gain_idx' and
returns '-INT_MAX', so it is easily to happen to place task on new CPU.

Furthermore, if review PE filter coordinate we also should consider
handling cases for coordinate axis. We can directly judge these the
cases for coordinate axis and origin into Suboptimal Region (SO) or
Optimal Region (O).  For instance, when the energy is same but the
capacity has less value, this should be (SO); for origin, we can take
it as (O).

This commit is to refine PE filter for coordinate axis and origin, so
the change follows the listed five conditions on X or Y axis:

  (nrg_delta, cap_delta) = (=0, <0)   --> Suboptimal Region
  (nrg_delta, cap_delta) = (=0, >0)   --> Optimal Region
  (nrg_delta, cap_delta) = (=0, =0)   --> Optimal Region
  (nrg_delta, cap_delta) = (<0, =0)   --> Optimal Region
  (nrg_delta, cap_delta) = (>0, =0)   --> Suboptimal Region

Change-Id: I553061276be64b800e519804da3fb6817db5d77d
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>